app/requirements.txt
apache-airflow>=3.1.8
lxml>=6.0.2
openpyxl>=3.1.5
python-jobspy>=1.1.82
